Renaming data package and formatting changes
