+++ nosejob/src/main/java/com/codingrangers/nosejob/models/data/parsing/ParsedProject.java
@@ -1,21 +1,20 @@
-package com.codingrangers.nosejob.models.data.parsed;
+package com.codingrangers.nosejob.models.data.parsing;
 
 import java.util.*;
 import com.codingrangers.nosejob.models.data.*;
 
 /**
- * ParsedProject
- * TODO: Need to unit test this
+ * ParsedProject TODO: Need to unit test this
  */
 public class ParsedProject implements ProjectData {
 
     private Map<String, ClassData> classes;
 
-    ParsedProject(){
+    ParsedProject() {
         classes = new HashMap<>();
     }
 
-    public void addClass(ClassData newClass){
+    public void addClass(ClassData newClass) {
         classes.put(newClass.getName(), newClass);
     }
 
@@ -27,5 +26,5 @@ public class ParsedProject implements ProjectData {
     @Override
     public ClassData getClassData(String className) {
         return classes.get(className);
-    } 
+    }
 }
